Add session and auth
Add web native auth. Auth is checked on page load. If a user is a guest, we
show them a sign in button. Otherwise, we don't show them a button. Later, we
will adding loading animations where needed.
Add Session, Viewer, and Api on the Elm side. A session can be logged in, loading,
or a guest. The viewer wraps the username of the authenticated user. The Api only
has auth for now, but we will do other port things here.
Add session to all pages. We hold the session in Shared, but sync it out when we
get one.