Add hex color parser and contrast determination
We can't use hex colors in the Elm app, but we can parse them to make Colors
that we can use. The parser handles all valid hex colors variants like #fff,
 #ffff, #ffffff, and #ffffffff.
Some theme and background colors work better with a black or white font, and
the contrast function figures out which one is best.