Add initial manifest load and create page save
On initial load of the app, we want to load up all of the users manifests. We
show the app, load the manifests, and update the manifest list when they are
ready.
The create page saves the manifest to web native and redirects to the homepage.
When the web native write completes, the list of manifests is updated with the
new manifest.